requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.1
attrs==25.4.0
bitsandbytes==0.45.5
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
executing==2.2.1
fastapi==0.121.2
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
hf-xet==1.2.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.36.0
idna==3.11
importlib_resources==6.5.2
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latex2mathml==3.78.1
llava==1.2.2.post1
markdown-it-py==4.0.0
markdown2==2.5.4
MarkupSafe==2.1.5
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.12.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.1.105
opencv-python==4.12.0.88
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
peft==0.5.0
pexpect==4.9.0
pillow==10.4.0
pip==25.3
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
ruff==0.14.5
safetensors==0.6.2
scikit-learn==1.2.2
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.1.99
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.49.3
svgwrite==1.4.3
sympy==1.14.0
threadpoolctl==3.6.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.37.2
triton==2.1.0
typer==0.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
visionzip==0.1.3
wavedrom==2.0.3.post3
wcwidth==0.2.14
websockets==11.0.3
wheel==0.45.1
